Refactor track classes into separate module
