+++ src/common/plug_indexes/generator.py
@@ -2,130 +2,11 @@
 import channels
 import plugins
 
-from .plugin import PluginIndex, AbstractTrack
-from typing import Literal, TypeVar, ParamSpec, Concatenate, Callable, Union
+from common.tracks import Channel
+from .plugin import PluginIndex
+from typing import Literal
 from common.types import Color
 from common.util.api_fixes import getGroupChannelIndex
-
-
-T = TypeVar('T')
-R = TypeVar('R')
-P = ParamSpec('P')
-
-
-def channelAction(
-    callback: Callable[Concatenate[int, P], T],
-    default_return: R,
-    global_index: int,
-    *args: P.args,
-    **kwargs: P.kwargs,
-) -> Union[T, R]:
-    """
-    Perform an action on a channel, only if it is possible to do so using
-    global indexes.
-
-    Because the FL Studio API doesn't support global indexes, this wrapper is
-    used to ensure
-
-    ### Args
-
-    * `callback`: function to call if possible - it should take a grouped index
-      as its first arg then, accept the remaining arguments given to this
-      function.
-
-    * `default_return` (`R`): return value if the callback cannot be safely
-      called.
-
-    * `global_index` (`int`): global index of channel to act on.
-
-    * `*args`, `**kwargs`: remaining args for callback
-
-    ### Returns
-
-    If the callback was called, a `tuple` of
-
-    * `True`: indicating success
-    * `T`: return value of `callback`
-
-    Or if it wasn't called, a `tuple` of
-
-    * `False`: indicating failure
-    * `R`: the given default return
-
-    ### Usage
-
-    ```py
-    success, value = channelAction(
-        channels.getChannelVolume,
-        'No volume',  # Default value
-        10,  # Channel has global index of 10
-        True,  # Get volume in decibels rather than as a percentage
-    )
-
-    if success:
-        reveal_type(value)  # float
-    else:
-        reveal_type(value)  # Literal['No volume']
-    ```
-    """
-    group_index = getGroupChannelIndex(global_index)
-
-    if group_index is None:
-        return default_return
-    else:
-        return callback(group_index, *args, **kwargs)
-
-
-class Channel(AbstractTrack):
-    """
-    Helper class for accessing properties of channels
-    """
-
-    def __init__(self, index: int) -> None:
-        self.__index = index
-
-    def triggerNote(self, note_number: int, value: float) -> None:
-        """
-        Trigger a note on this channel
-        """
-        channels.midiNoteOn(self.__index, note_number, int(value * 127))
-
-    @property
-    def color(self) -> Color:
-        """
-        Color of the channel
-        """
-        return Color.fromInteger(channelAction(
-            channels.getChannelColor,
-            0,
-            self.__index
-        ))
-
-    @color.setter
-    def color(self, new_color: Color) -> None:
-        channelAction(
-            channels.setChannelColor,
-            None,
-            self.__index,
-            new_color.integer,
-        )
-
-    @property
-    def name(self) -> str:
-        return channelAction(
-            channels.getChannelName,
-            '',
-            self.__index
-        )
-
-    @name.setter
-    def name(self, new_name: str) -> None:
-        channelAction(
-            channels.setChannelName,
-            None,
-            self.__index,
-            new_name,
-        )
 
 
 class GeneratorIndex(PluginIndex):
